Split event log writer from scenario runner
